Fixed issue 139
saveSettings.tcl - Fixed an issue where the Header variables would be over written with blank data when saved, if we didn't go to that page.

setup_carrierMethod_gui.tcl - Fixed the width of the listbox widgets.
setup_packaging_gui.tcl - Fixed the width of the listbox widgets.

startup.tcl - Commented out the function which tries to look for updates to Efficiency Assist. This needs more work.